Two sets of changes
1) am now allocating initial trie hash-table array out of a
   structure manager rather than via mallocs, at least for the MT
   engine.

2) Main changes are to the C interface with the MT engine.  I'm
   probably not done with these changes, but by putting them into
   CVS it allows Rui and I to test on different versions of Linux.

-- Dipti's bug not yet fixed.  I wanted to get these changes out
   of the way first.

Terry

+++ emu/struct_manager.c
@@ -66,7 +66,7 @@ void smPrint(Structure_Manager smRecord, char *string) {
 	  SM_CurBlock(smRecord),	nBlocks,
 	  SM_NextStruct(smRecord),	SM_FreeList(smRecord),
 	  SM_LastStruct(smRecord),	SM_AllocList(smRecord),
-	  SM_StructsPerBlock(smRecord),	SM_StructSize(smRecord));
+	  SM_StructsPerBlock(smRecord),	(int) SM_StructSize(smRecord));
 }
 
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
